refactor: structure of repository in leaderboard-scoring app and add in memory queue repo to repository layer
