+++ deploy/leaderboardscoring/development/config.yml
@@ -57,6 +57,6 @@ consumer:
 
 total_shutdown_timeout: 30s
 
-path_of_migration: "./leaderboardscoringapp/repository/migrations"
+path_of_migration: "./leaderboardscoringapp/repository/database/migrations"
 
 subscriber_topic: "CONTRIBUTION_REGISTERED"
