+++ leaderboardscoringapp/repository/database/database_repository.go
@@ -1,0 +1,11 @@
+package database
+
+import "github.com/jackc/pgx/v5/pgxpool"
+
+type Repository struct {
+	db *pgxpool.Pool
+}
+
+func New(db *pgxpool.Pool) Repository {
+	return Repository{}
+}
