+++ leaderboardscoringapp/repository/dbconfig.yml
@@ -6,4 +6,4 @@ postgres_db:
   db_name: "leaderboardscoring_db"
   ssl_mode: "disable"
 
-path_of_migration: "./leaderboardscoring/repository/migrations"
+path_of_migration: "./leaderboardscoring/repository/database/migrations"
