+++ leaderboardscoringapp/repository/inmemory/inmemory_queue_repository.go
@@ -1,0 +1,11 @@
+package inmemoryrepository
+
+import "github.com/patrickmn/go-cache"
+
+type InMemoryBuffered struct {
+	c *cache.Cache
+}
+
+func New() InMemoryBuffered {
+	return InMemoryBuffered{}
+}
